+++ src/index.js
@@ -6,6 +6,7 @@ import getRankChange from "./rank-change";
 import getTeamIcons from "./team-icons";
 import initShishenRank from "./init/shishen-rank";
 import initTables from "./init/table";
+import initOptions from "./init/shishen";
 
 $(window).on('load', function () {
     console.log("window done");
